Rename files P1 (read description)
The 'weird bug' which there are now several commits fixing is caused by the fact that for some reason Git does not detect when you rename a file if the only thing you changed was capitalization. Well I had renamed some things to be capitalized during a larger commit, and didn't notice they never went through. This caused a mismatch in the cases between remote and my machine, which caused some random annoying errors every time I merged with main.

Anyway, this is part 1 because I had to change the text in some way other than just capitalization in order for git to notice. The next commit will be removing the extra text.